Refactored config quality
