+++ tsdat/config/quality/checker_config.py
@@ -1,0 +1,5 @@
+from ..utils import ParameterizedConfigClass
+
+
+class CheckerConfig(ParameterizedConfigClass):
+    pass
