+++ tsdat/config/quality/handler_config.py
@@ -1,0 +1,5 @@
+from ..utils import ParameterizedConfigClass
+
+
+class HandlerConfig(ParameterizedConfigClass):
+    pass
